Refactor: extract formatDuration utility and optimize chart rendering with useMemo
