+++ src/webview/coverageView/coverageHeader.tsx
@@ -1,6 +1,8 @@
 import * as React from 'react';
+import { VSCodeButton } from '@vscode/webview-ui-toolkit/react';
 import { CoverageData } from '../../model/coverage';
 import { vscode } from './vscode';
+import { formatDuration } from '../common/formatters';
 import './coverageHeader.css';
 
 interface CoverageHeaderProps {
@@ -13,19 +15,6 @@ export const CoverageHeader: React.FC<CoverageHeaderProps> = ({ data }) => {
     };
 
     const lastStat = data.stats[data.stats.length - 1];
-    const formatDuration = (seconds: number): string => {
-        const hours = Math.floor(seconds / 3600);
-        const minutes = Math.floor((seconds % 3600) / 60);
-        const secs = seconds % 60;
-
-        if (hours > 0) {
-            return `${hours}h ${minutes}m ${secs}s`;
-        } else if (minutes > 0) {
-            return `${minutes}m ${secs}s`;
-        } else {
-            return `${secs}s`;
-        }
-    };
 
     return (
         <div className="coverage-header">
@@ -59,9 +48,9 @@ export const CoverageHeader: React.FC<CoverageHeaderProps> = ({ data }) => {
                     </div>
                 </div>
                 <div className="header-actions">
-                    <vscode-button appearance="icon" onClick={handleRefresh}>
+                    <VSCodeButton appearance="icon" onClick={handleRefresh}>
                         <span className="codicon codicon-refresh"></span>
-                    </vscode-button>
+                    </VSCodeButton>
                 </div>
             </div>
         </div>
